+++ iterset.go
@@ -79,12 +79,12 @@ func (m MapSet[K, V]) IsSubset(keys iter.Seq[K]) bool {
 	return len(m) == len(s)
 }
 
-// IsSubset returns whether all keys in seq1 are in seq2.
-// Note [MapSet.IsSuperset] is more efficient if seq2 is from a map.
+// IsSubset returns whether all keys are present in the sequence.
+// Note [MapSet.IsSuperset] is more efficient if the sequence is from a map.
 //   - time: Θ(k)
 //   - space: Θ(k)
-func IsSubset[K comparable](seq1, seq2 iter.Seq[K]) bool {
-	return Collect(seq2, struct{}{}).IsSuperset(seq1)
+func IsSubset[K comparable](keys, seq iter.Seq[K]) bool {
+	return Collect(seq, struct{}{}).IsSuperset(keys)
 }
 
 // IsSuperset returns whether all keys are present.
@@ -145,6 +145,17 @@ func (m MapSet[K, V]) Intersect(keys iter.Seq[K]) iter.Seq2[K, V] {
 	}
 }
 
+// Intersect returns the ordered keys which are present in the sequence(s).
+// Note [MapSet.Intersect] is more efficient if the sequence is from a map.
+//   - time: Θ(k)
+//   - space: Θ(k)
+func Intersect[K comparable](keys iter.Seq[K], seqs ...iter.Seq[K]) iter.Seq[K] {
+	for _, seq := range seqs {
+		keys = filterFunc(keys, Collect(seq, struct{}{}).contains)
+	}
+	return keys
+}
+
 // Difference returns the key-value pairs which are not present in the keys.
 // Note [MapSet.ReverseDifference] is more efficient if the keys are from a map.
 // [Difference] is more efficient if the receiver was not originally a map.
@@ -164,12 +175,15 @@ func (m MapSet[K, V]) Difference(keys iter.Seq[K]) iter.Seq2[K, V] {
 	}
 }
 
-// Difference returns keys in seq1 that are not in seq2.
-// Note [MapSet.ReverseDifference] is more efficient if seq2 is from a map.
+// Difference returns the ordered keys which are not present in the sequence(s).
+// Note [MapSet.ReverseDifference] is more efficient if the sequence is from a map.
 //   - time: Θ(k)
 //   - space: Θ(k)
-func Difference[K comparable](seq1, seq2 iter.Seq[K]) iter.Seq[K] {
-	return Collect(seq2, struct{}{}).ReverseDifference(seq1)
+func Difference[K comparable](keys iter.Seq[K], seqs ...iter.Seq[K]) iter.Seq[K] {
+	for _, seq := range seqs {
+		keys = filterFunc(keys, Collect(seq, struct{}{}).missing)
+	}
+	return keys
 }
 
 // ReverseDifference returns the ordered keys which are not present in the map.