+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 ## Unreleased
+## Added
+* `Intersect` function
 
 ## [0.2.0](https://github.com/coady/iterset/releases/tag/v0.2.0) - 2024-12-06
 ### Added